backend/requirements.txt
# Web Framework
flask==2.3.0
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-limiter==3.3.0

# Database
sqlalchemy==2.0.0
psycopg2-binary==2.9.6
alembic==1.11.0

# Data Processing
pandas==2.0.0
numpy==1.24.0

# Machine Learning
scikit-learn==1.3.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==4.5.0

# Testing
pytest==7.3.0
pytest-cov==4.1.0
pytest-flask==1.2.0

# Code Quality
flake8==6.0.0
black==23.3.0

# Optional - FastAPI alternative
# fastapi==0.95.0
# uvicorn[standard]==0.22.0
# pydantic==1.10.8
